refactor(ably-subscriber): share example token request signing
