+++ crates/ably-subscriber/examples/subscribe.rs
@@ -13,47 +13,9 @@
 //! `API_KEY` format: `keyName:keySecret` (from your Ably dashboard).
 //! Message data is printed to stdout (pipe to `jq` for formatting).
 
-use std::time::{SystemTime, UNIX_EPOCH};
+use ably_subscriber::{Event, SubscribeConfig, subscribe};
 
-use base64::Engine as _;
-use base64::engine::general_purpose::STANDARD as BASE64;
-use hmac::{Hmac, KeyInit, Mac};
-use sha2::Sha256;
-
-use ably_subscriber::{Event, SubscribeConfig, TokenRequest, subscribe};
-
-type HmacSha256 = Hmac<Sha256>;
-
-fn create_token_request(
-    key_name: &str,
-    key_secret: &str,
-) -> Result<TokenRequest, Box<dyn std::error::Error + Send + Sync>> {
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map_err(|e| format!("system time error: {e}"))?;
-    let timestamp = now.as_millis() as i64;
-    let nonce = format!("{:x}{:x}", now.as_nanos(), std::process::id());
-    let ttl = 3_600_000_i64; // 1 hour
-    let capability = r#"{"*":["*"]}"#;
-
-    // Ably signing format: keyName\nttl\ncapability\nclientId\ntimestamp\nnonce\n
-    let sign_text = format!("{key_name}\n{ttl}\n{capability}\n\n{timestamp}\n{nonce}\n");
-
-    let mut mac = HmacSha256::new_from_slice(key_secret.as_bytes())
-        .map_err(|e| format!("HMAC error: {e}"))?;
-    mac.update(sign_text.as_bytes());
-    let mac_b64 = BASE64.encode(mac.finalize().into_bytes());
-
-    Ok(TokenRequest {
-        key_name: key_name.to_string(),
-        timestamp,
-        nonce,
-        mac: mac_b64,
-        capability: capability.to_string(),
-        ttl: Some(ttl),
-        client_id: None,
-    })
-}
+mod common;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -88,7 +50,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         Box::new(move || {
             let kn = key_name.clone();
             let ks = key_secret.clone();
-            Box::pin(async move { create_token_request(&kn, &ks) })
+            Box::pin(async move { common::create_token_request(&kn, &ks, common::ONE_HOUR_TTL_MS) })
         }),
         channel.to_string(),
     );
